Fork a thread for provisioning via web & rest
To make the interface responsive, a child thread is forked to
handle the provisioning. This means that I need to come up with a good
way to get messages back to the client.

I may switch the REST call back to blocking. Or maybe I'll make that
selectable. I'm not sure yet. The GUI cannot wait for the provisioning
to complete though. Perhaps the best solution is to log actions and
provide a log tab.